# backend/public/openapi.yaml
openapi: 3.1.0
info:
  title: Cultural Site API
  version: 1.0.0
  description: API for managing cultural sites, proposals, reviews, and user data.
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CulturalSite:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the cultural site.
        name:
          type: string
          description: Name of the cultural site.
          minLength: 2
          maxLength: 100
        description:
          type: string
          description: Description of the cultural site.
          maxLength: 1000
        category:
          type: string
          description: Category of the cultural site.
          enum:
            - artwork
            - gallery
            - museum
            - restaurant
            - theatre
            - arts_centre
            - community_centre
            - library
            - cinema
            - other
        location:
          type: object
          properties:
            type:
              type: string
              enum: ['Point']
              description: GeoJSON type.
            coordinates:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
              description: |
                [longitude, latitude] coordinates. Longitude must be between -180 and 180,
                latitude must be between -90 and 90.
          required:
            - type
            - coordinates
        address:
          type: string
          description: Address of the cultural site.
        website:
          type: string
          format: url
          description: Website URL of the cultural site.
        imageUrl:
          type: string
          format: url
          description: Image URL of the cultural site.
        openingHours:
          type: string
          description: Opening hours of the cultural site.
        licenseInfo:
          type: string
          default: 'Data © OpenStreetMap contributors, ODbL.'
          description: Licensing information for the data.
        sourceId:
          type: string
          description: OSM ID (node/way/relation ID) from Overpass API. (node/1234567890)
        reviews:
          type: array
          items:
            type: string
            description: ObjectId of a Review.
        favoritesCount:
          type: number
          format: int32
          minimum: 0
          default: 0
          description: Number of users who favorited this site.
        originalTags:
          type: object
          description: Original GeoJSON 'tags' object from the source element from Overpass API response.
          additionalProperties: true
        proposedBy:
          type: string
          description: ObjectId of the User who proposed this site.
        registeredBy:
          type: string
          description: ObjectId of the User (admin) who registered this site.
        active:
          type: boolean
          default: true
          description: Field for soft deletion (not currently implemented).
      required:
        - name
        - category
        - location
        - address
        - sourceId
        - originalTags
...
paths:
...
  /cultural-sites/nearby-osm:
    get:
      summary: Get nearby cultural sites from OSM database for a lon/lat coordinates
      tags:
        - Cultural Sites
      operationId: getNearbyOsmCulturalSites
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: latitude
          schema:
            type: number
            format: float
          required: true
          description: Latitude for the center of the search.
        - in: query
          name: longitude
          schema:
            type: number
            format: float
          required: true
          description: Longitude for the center of the search.
        - in: query
          name: radius
          schema:
            type: number
            format: int32
            default: 1000 # Corrected: default is inside schema
          description: Search radius in meters.
      responses:
        200:
          description: A list of nearby cultural sites from OSM.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CulturalSite'
        400:
          description: Invalid latitude/longitude or location not within Chemnitz.
        401:
          description: Unauthorized if no token is provided.
        403:
          description: Forbidden if user is not 'admin' or 'user'.
        500:
          description: Internal server error.
  /cultural-sites:
    get:
      summary: Get all cultural sites
      tags:
        - Cultural Sites
      operationId: getAllCulturalSites
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: Filter by cultural site category.
        - in: query
          name: name
          schema:
            type: string
          description: Filter by cultural site name (partial match).
      responses:
        200:
          description: A list of cultural sites.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CulturalSite'
        500:
          description: Internal server error.
    post:
      summary: Save a cultural site to the database
      tags:
        - Cultural Sites
      operationId: saveCulturalSiteToDb
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CulturalSite'
            examples:
              newCulturalSite:
                value:
                  name: 'Example Museum'
                  description: 'A wonderful museum of modern art.'
                  category: 'Museum'
                  location:
                    type: 'Point'
                    coordinates: [12.345, 54.321]
                  address: '123 Museum St, City'
                  website: 'http://examplemuseum.com'
                  sourceId: 'osm_node_12345'
                  originalTags:
                    amenity: 'museum'
                    name: 'Example Museum'
      responses:
        201:
          description: Cultural site successfully saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CulturalSite'
        400:
          description: Invalid input (e.g., missing fields, invalid location, not in Chemnitz).
        401:
          description: Unauthorized.
        403:
          description: Forbidden (only admins can add sites, or sourceId is in exclusion list).
        409:
          description: Conflict (e.g., sourceId already registered).
        500:
          description: Internal server error during saving.
  /cultural-sites/{id}:
    get:
      summary: Get a cultural site by ID
      tags:
        - Cultural Sites
      operationId: getCulturalSiteById
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Unique ID of the cultural site.
      responses:
        200:
          description: Details of the cultural site.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CulturalSite'
        400:
          description: Invalid cultural site ID format.
        404:
          description: Cultural site not found.
        500:
          description: Internal server error.
    put:
      summary: Update a cultural site by ID (Admin only)
      tags:
        - Cultural Sites
      operationId: updateCulturalSiteById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Unique ID of the cultural site.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CulturalSite'
            examples:
              updateCulturalSite:
                value:
                  description: 'Updated description of the museum.'
                  openingHours: '9 AM - 5 PM'
      responses:
        200:
          description: Cultural site updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CulturalSite'
        400:
          description: Invalid input.
        401:
          description: Unauthorized.
        403:
          description: Forbidden, only admins can update sites.
        404:
          description: Cultural site not found.
    delete:
      summary: Delete a cultural site by ID (Admin only)
      tags:
        - Cultural Sites
      operationId: deleteCulturalSiteById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Unique ID of the cultural site.
      responses:
        204:
          description: Cultural site successfully deleted.
        401:
          description: Unauthorized.
        403:
          description: Forbidden (only admins can delete sites).
        404:
          description: Cultural site not found.
        409:
          description: Conflict (e.g., sourceId already exists in exclusion list).
        500:
          description: Internal server error during deletion.
...